Change float to BigDecimal

+++ products/src/main/java/com/inventory/products/dto/CategoryMetrics.java
@@ -3,11 +3,13 @@ package com.inventory.products.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class CategoryMetrics {
     private String categoryName;
     private int totalProductsInStock;
-    private double totalValueInStock;
-    private double averagePriceInStock;
+    private BigDecimal totalValueInStock;
+    private BigDecimal averagePriceInStock;
 }
