+++ products/src/main/java/com/inventory/products/dto/OverallMetrics.java
@@ -3,10 +3,12 @@ package com.inventory.products.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class OverallMetrics {
     private int totalProductsInStock;
-    private double totalValueInStock;
-    private double averagePriceInStock;
+    private BigDecimal totalValueInStock;
+    private BigDecimal averagePriceInStock;
 }
